fix: change method defs to 8 bit WAV depth

+++ lib/CarlAudio/src/WAVFile.hpp
@@ -14,7 +14,7 @@
  */
 class WAVFile {
   private:
-    uint16_t* samples; /**< the samples cached for this file */
+    uint8_t* samples; /**< the samples cached for this file */
     int16_t number_samples; /**< the number of samples cached for this file */
 
   public:
@@ -30,7 +30,7 @@ class WAVFile {
      * @param position which sample to get
      * @return the sample
      */
-    uint16_t get_sample(int16_t position);
+    uint8_t get_sample(int16_t position);
     /**
      * @brief get the number of samples
      * @return the number of samples
